Refactor admin handlers into separate modules
