+++ main.py
@@ -14,7 +14,7 @@ from database.engine import create_db, drop_db, session_maker
 
 from handlers.user_private import user_private_router
 from handlers.user_group import user_group_router
-from handlers.admin_private import admin_router
+from handlers.admin_hendlers import admin_router
 from handlers.order_processing import order_router
 
 # ALLOWED_UPDATES = ['message', 'edited_message', 'callback_query']
